Update README.md
A supplementary material containing the data and model files used in the manuscript "Density perturbations in the crust indicate potential for blind magmatism beneath magma-poor rifts".  The supplementary material contains the processed mantle corrected gravity data and the netCDF format (gravdata.nc). The main density model from jif3D is result.inv.grav.nc. We have also included text files of the extracted density perturbations at depths of 15, 20, 50, and 100 km. Please take a look at the manuscript and the files included for more information on the origin of the data. There is also a text file of the observed versus the predicted gravity data. A link will be added upon acceptance.